fix(themes): use README feed and video for discussion previews
