+++ e2e/screenshots/capture.spec.mjs
@@ -1,11 +1,8 @@
 import path from 'node:path'
 
-import { test, expect, goTo, repoRoot, sel } from '../helpers/app.mjs'
-import { captureAppScreenshot, copyScreenshots, openScreenshotSettings, sizeScreenshotWindow } from '../helpers/screenshots.mjs'
+import { test, expect, repoRoot } from '../helpers/app.mjs'
+import { captureAppScreenshot, copyScreenshots, openScreenshotSettings, openScreenshotSubscriptions, openScreenshotWatch, sizeScreenshotWindow } from '../helpers/screenshots.mjs'
 
-// Keep both themes on the same video and decoded frame across future updates.
-const VIDEO_ID = 'AY5qcIq5u2g'
-const TIMESTAMP = 3 * 3600 + 41 * 60 + 58
 const THEMES = ['dark', 'light']
 
 test.use({
@@ -38,15 +35,6 @@ async function setTheme(app, theme) {
   await expect(app.page.locator('body')).toHaveClass(new RegExp(`\\b${theme}\\b`))
 }
 
-async function expectPausedFrame(video) {
-  await expect.poll(() => video.evaluate(element => ({
-    paused: element.paused,
-    seeking: element.seeking,
-    ready: element.readyState >= HTMLMediaElement.HAVE_CURRENT_DATA,
-    time: element.currentTime,
-  }))).toEqual({ paused: true, seeking: false, ready: true, time: TIMESTAMP })
-}
-
 test('refresh README screenshots from the live app', async ({ app, page }, testInfo) => {
   const tutorial = page.locator('.tutorialOverlay')
   await tutorial.getByRole('button', { name: 'Skip', exact: true }).click()
@@ -73,27 +61,8 @@ test('refresh README screenshots from the live app', async ({ app, page }, testI
     await expect(dialog).toBeHidden()
   })
 
-  // Configure only what the live content capture needs, after photographing
-  // Settings with a fresh profile and all app defaults intact.
-  await page.evaluate(async () => {
-    const store = document.querySelector('#app').__vue_app__.config.globalProperties.$store
-    await store.dispatch('updateFetchSubscriptionsAutomatically', false)
-    await store.dispatch('updateVideoPlaybackEngine', 'yt-dlp')
-  })
-
   await test.step('subscriptions in both themes', async () => {
-    await page.locator(sel.searchInput).fill('https://www.youtube.com/channel/UCsXVk37bltHxD1rDPwtNM8Q')
-    await page.locator(sel.searchInput).press('Enter')
-    const subscribe = page.locator('.ftSubscribeButton .subscribeButton').first()
-    await expect(subscribe).toHaveText(/^\s*Subscribe/)
-    await subscribe.click()
-    await expect(subscribe).toHaveText(/^\s*Unsubscribe/)
-    await page.locator(sel.searchInput).fill('')
-    await goTo(page, 'subscriptions')
-    await page.locator('.subscriptionsHeaderRefreshWidget .refreshButton').click()
-    await expect(page.locator('.ft-list-video').nth(11)).toBeVisible({ timeout: 60_000 })
-    await expect(page.locator('.subscriptionsHeaderRefreshWidget .refreshButton button'))
-      .toHaveAttribute('title', /Refresh/)
+    await openScreenshotSubscriptions(page)
     for (const theme of THEMES) {
       await setTheme(app, theme)
       await capture(1, theme)
@@ -101,27 +70,10 @@ test('refresh README screenshots from the live app', async ({ app, page }, testI
   })
 
   await test.step('watch page at 3:41:58 in both themes', async () => {
-    await page.locator(sel.searchInput).fill(`https://www.youtube.com/watch?v=${VIDEO_ID}&t=${TIMESTAMP}`)
-    await page.locator(sel.searchInput).press('Enter')
-    await expect(page.locator('.videoTitle')).toContainText('FLYING OVER JAPAN', { timeout: 60_000 })
-    const video = page.locator('.ftVideoPlayer video')
-    await expect(video).toBeVisible({ timeout: 60_000 })
-    await expect.poll(async () => ({
-      ready: await video.evaluate(element => element.readyState >= 2),
-      errors: await page.locator('.errorMessage:visible').allTextContents(),
-    }), { timeout: 60_000, message: 'waiting for a playable frame from YouTube' })
-      .toEqual({ ready: true, errors: [] })
-    await video.evaluate((element, timestamp) => {
-      element.pause()
-      element.currentTime = timestamp
-    }, TIMESTAMP)
-    await expectPausedFrame(video)
-    await page.locator(sel.searchInput).fill('')
+    await openScreenshotWatch(page)
     for (const theme of THEMES) {
       await setTheme(app, theme)
-      await expectPausedFrame(video)
       await capture(2, theme)
-      await expectPausedFrame(video)
     }
   })
